Refactor login page components and file structure
