+++ DREAM Lab Frontend/src/Login/LoginPage/LoginPage.css
@@ -15,25 +15,23 @@
     height: 70vh;
   }
   @media screen and (min-width: 430px) {
-    width: 400px;
+    width: 430px;
     height: 80vh;
+  }
+  @media screen and (min-height: 800px) {
+    width: 430px;
+    height: 700px;
   }
 }
 
-.login-logo-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .login-logo {
-  height: 30px;
+  height: 40px;
   margin-right: 15px;
 }
 
 .login-dreamlab-title {
   color: white;
-  font-size: 20px;
+  font-size: 25px;
   font-weight: bold;
 }
 
@@ -41,10 +39,4 @@
   color: white;
   font-size: 28px;
   text-align: center;
-}
-
-.login-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
 }
